chore: rename useEvent hook
